General view for both editing and adding finished.

+++ src/DogShelter/Model/Dog.cs
@@ -7,7 +7,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public bool Male { get; set; }
-        public uint Age { get; set; }
+        public double Age { get; set; }
         public DateTime BroughtToShelter { get; set; }
         public string Description { get; set; }
         public string CareTaker { get; set; }
